When level is unknown, don't go up but wait for a first setLevel command
Allow LEVEL_NONE as valid in the StoredState.
At initialization, don't go up immediately, but wait for a setLevel
command that is 0 or 100 to start moving.
Only then, go to the STATE_RESETTING state.
When in STATE_RESETTING state, don't accept other setLevel or
stop commands until the move is finished, that is we are back in
STATE_IDLE with a known level.